Check if KAFKA_HOME is set before running the script

+++ Tools/src/main/startIasServices.sh
@@ -1,5 +1,9 @@
 #!/bin/bash 
-
+if [ -z $KAFKA_HOME ]
+then
+	echo "Set KAFKA_HOME before running this script"
+	exit 1
+fi
 echo "Cleaning folders..."
 rm -rf /opt/kafkadata/* /opt/zookeeperdata/*
 cd $KAFKA_HOME
